refactor: apply PR review follow-ups to jdbc repositories and instructor tests
